Moving JWT Key responsibilities from database to API
