+++ src/api/app_factory.py
@@ -7,27 +7,33 @@ from src.database.db import db
 from src.database.users.models import BlacklistedToken
 
 from src.database.config_db import Config, initialize_database
+from src.api.config.config import APIConfig
 
 
 def create_app():
     app = Flask(__name__)
+    
+    # Load database configuration
     app.config.from_object(Config)
+    
+    # Load API configuration (JWT, CORS, etc.)
+    app.config['JWT_SECRET_KEY'] = APIConfig.JWT_SECRET_KEY
+    app.config['JWT_ACCESS_TOKEN_EXPIRES'] = APIConfig.JWT_ACCESS_TOKEN_EXPIRES
+    app.config['JWT_TOKEN_LOCATION'] = APIConfig.JWT_TOKEN_LOCATION
 
     # Configure CORS to allow requests from the frontend
     CORS(app, 
         resources={r"/*": {
-            "origins": ["http://localhost:5000", "http://127.0.0.1:5000"],
-            "methods": ["GET", "HEAD", "POST", "OPTIONS", "PUT", "PATCH", "DELETE"],
-            "allow_headers": ["Content-Type", "Authorization", "X-Requested-With"],
-            "supports_credentials": True
+            "origins": APIConfig.CORS_ORIGINS,
+            "methods": APIConfig.CORS_METHODS,
+            "allow_headers": APIConfig.CORS_ALLOW_HEADERS,
+            "supports_credentials": APIConfig.CORS_SUPPORTS_CREDENTIALS
         }})
 
     # initialize extensions
     db.init_app(app)
 
-    # Add JWT configuration
-    app.config['JWT_SECRET_KEY'] = Config.JWT_SECRET_KEY
-    app.config['JWT_TOKEN_LOCATION'] = ['headers', 'cookies']
+    # Initialize JWT
     jwt = JWTManager(app)
 
     @jwt.token_in_blocklist_loader
