+++ src/database/config_db.py
@@ -1,16 +1,14 @@
 import os
 from sqlalchemy import inspect
 from src.database.db import db
-from datetime import timedelta
 from dotenv import load_dotenv
 from pathlib import Path
 from flask import Flask
 import logging
 
-
-
-config_dir = Path(__file__).parent
-load_dotenv(os.path.join(config_dir, '.env'))
+# Note: .env file is now in src/api/config/ 
+# This config only handles database settings
+# For JWT and API settings, see src/api/config/config.py
 
 class Config:
     BASE_DIR = Path(__file__).resolve().parent.parent.parent
@@ -21,22 +19,6 @@ class Config:
         f"sqlite:///{DB_PATH}"
     )
     SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-    # Load JWT secret key from environment
-    try:
-        JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
-        if not JWT_SECRET_KEY:
-            raise ValueError("JWT_SECRET_KEY not found in environment variables")
-    except Exception as e:
-        logging.error("JWT_SECRET_KEY must be set in .env file")
-        logging.error("Please create a .env file in src/config with JWT_SECRET_KEY=your_secret_key")
-        raise e
-
-    JWT_ACCESS_TOKEN_EXPIRES = timedelta(hours=12)
-
-    @staticmethod
-    def get_token_expires_delta():
-        return timedelta(hours=12)
 
 
 def initialize_database(app):
